validation for empty input fix
